feat: add calendar and event components, refactor imports for better organization
