+++ src/components/calendar/MonthView.tsx
@@ -10,8 +10,8 @@ import {
   TableRow,
   Typography,
 } from '@mui/material';
-import { Event } from '../types';
-import { formatDate, formatMonth, getEventsForDay, getWeeksAtMonth } from '../utils/dateUtils';
+import { Event } from '../../types';
+import { formatDate, formatMonth, getEventsForDay, getWeeksAtMonth } from '../../utils/dateUtils';
 
 const weekDays = ['일', '월', '화', '수', '목', '금', '토'];
