+++ src/components/form/EventForm.tsx
@@ -10,8 +10,8 @@ import {
   Tooltip,
   Typography,
 } from '@mui/material';
-import { EventForm as EventFormType } from '../types';
-import { getTimeErrorMessage } from '../utils/timeValidation';
+import { EventForm as EventFormType } from '../../types';
+import { getTimeErrorMessage } from '../../utils/timeValidation';
 
 const categories = ['업무', '개인', '가족', '기타'];
