+++ src/components/index.ts
@@ -1,0 +1,11 @@
+// Calendar components
+export * from './calendar';
+
+// Event components
+export * from './event';
+
+// Form components
+export * from './form';
+
+// Notification components
+export * from './notification';
