+++ src/components/notification/index.ts
@@ -1,0 +1,1 @@
+export { NotificationStack } from './NotificationStack';
